agents/requirements.txt
python-dotenv==1.0.1
openai==1.68.2
requests==2.32.3